Mode the debug console to DebugConsole
