+++ main/devices/Device.hpp
@@ -13,10 +13,10 @@
 #include <kernel/BootClock.hpp>
 #include <kernel/Command.hpp>
 #include <kernel/Concurrent.hpp>
+#include <kernel/DebugConsole.hpp>
 #include <kernel/Kernel.hpp>
 #include <kernel/Strings.hpp>
 #include <kernel/Task.hpp>
-#include <kernel/drivers/RtcDriver.hpp>
 #include <kernel/mqtt/MqttDriver.hpp>
 #include <kernel/mqtt/MqttRoot.hpp>
 
@@ -51,98 +51,6 @@ typedef farmhub::devices::Mk8Config TDeviceConfiguration;
 #endif
 
 namespace farmhub::devices {
-
-#ifdef FARMHUB_DEBUG
-class ConsolePrinter {
-public:
-    ConsolePrinter(const std::shared_ptr<BatteryManager> battery, WiFiDriver& wifi)
-        : battery(battery)
-        , wifi(wifi) {
-        status.reserve(256);
-        Task::loop("console", 3072, 1, [this](Task& task) {
-            printStatus();
-            task.delayUntilAtLeast(250ms);
-        });
-    }
-
-private:
-    void printStatus() {
-        static const char* spinner = "|/-\\";
-        static const int spinnerLength = strlen(spinner);
-        auto uptime = duration_cast<milliseconds>(boot_clock::now().time_since_epoch());
-
-        counter = (counter + 1) % spinnerLength;
-        status.clear();
-        status += "[" + std::string(1, spinner[counter]) + "] ";
-        status += "\033[33m" + std::string(farmhubVersion) + "\033[0m";
-        status += ", uptime: \033[33m" + toStringWithPrecision(uptime.count() / 1000.0f, 1) + "\033[0m s";
-        status += ", WIFI: " + std::string(wifiStatus()) + " (up \033[33m" + toStringWithPrecision(wifi.getUptime().count() / 1000.0f, 1) + "\033[0m s)";
-        status += ", RTC \033[33m" + std::string(RtcDriver::isTimeSet() ? "OK" : "UNSYNCED") + "\033[0m";
-        status += ", heap \033[33m" + toStringWithPrecision(heap_caps_get_free_size(MALLOC_CAP_INTERNAL) / 1024.0f, 2) + "\033[0m kB";
-        status += ", CPU: \033[33m" + std::to_string(esp_clk_cpu_freq() / 1000000) + "\033[0m MHz";
-
-        if (battery != nullptr) {
-            status += ", battery: \033[33m" + toStringWithPrecision(battery->getVoltage(), 2) + "\033[0m V";
-        }
-
-        printf("\033[1G\033[0K%s", status.c_str());
-        fflush(stdout);
-        fsync(fileno(stdout));
-    }
-
-    static const char* wifiStatus() {
-        auto netif = esp_netif_get_default_netif();
-        if (!netif) {
-            return "\033[0;33moff\033[0m";
-        }
-
-        wifi_mode_t mode;
-        ESP_ERROR_CHECK(esp_wifi_get_mode(&mode));
-
-        switch (mode) {
-            case WIFI_MODE_STA:
-                break;
-            case WIFI_MODE_NULL:
-                return "\033[0;33mNULL\033[0m";
-            case WIFI_MODE_AP:
-                return "\033[0;32mAP\033[0m";
-            case WIFI_MODE_APSTA:
-                return "\033[0;32mAPSTA\033[0m";
-            case WIFI_MODE_NAN:
-                return "\033[0;32mNAN\033[0m";
-            default:
-                return "\033[0;31m???\033[0m";
-        }
-
-        // Retrieve the current Wi-Fi station connection status
-        wifi_ap_record_t ap_info;
-        esp_err_t err = esp_wifi_sta_get_ap_info(&ap_info);
-        if (err != ESP_OK) {
-            return esp_err_to_name(err);
-        }
-
-        // Check IP address
-        esp_netif_ip_info_t ip_info;
-        err = esp_netif_get_ip_info(netif, &ip_info);
-        if (err != ESP_OK) {
-            return esp_err_to_name(err);
-        }
-
-        if (ip_info.ip.addr != 0) {
-            static char ip_str[32];
-            snprintf(ip_str, sizeof(ip_str), "\033[0;33m" IPSTR "\033[0m", IP2STR(&ip_info.ip));
-            return ip_str;
-        } else {
-            return "\033[0;33mIP?\033[0m";
-        }
-    }
-
-    int counter;
-    std::string status;
-    const std::shared_ptr<BatteryManager> battery;
-    WiFiDriver& wifi;
-};
-#endif
 
 class MemoryTelemetryProvider : public TelemetryProvider {
 public:
@@ -211,7 +119,7 @@ public:
         , deviceDefinition(deviceDefinition)
         , kernel(kernel)
 #ifdef FARMHUB_DEBUG
-        , consolePrinter(battery, kernel->wifi)
+        , debugConsole(battery, kernel->wifi)
 #endif
     {
         kernel->switches->onReleased("factory-reset", deviceDefinition->bootPin, SwitchMode::PullUp, [this](const Switch&, milliseconds duration) {
@@ -471,7 +379,7 @@ private:
     const std::shared_ptr<Kernel> kernel;
 
 #ifdef FARMHUB_DEBUG
-    ConsolePrinter consolePrinter;
+    DebugConsole debugConsole;
 #endif
 
     std::shared_ptr<MqttRoot> mqttDeviceRoot = kernel->mqtt->forRoot(locationPrefix() + "devices/ugly-duckling/" + instance);
